HUD overlay refactors - Rename to IGuiOverlay match vanilla (instead of Ingame) - Add overlay manager and registration event - Move vanilla overlays to a standalone enum
