+++ src/main/java/net/minecraftforge/client/gui/overlay/NamedGuiOverlay.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Forge Development LLC and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
+package net.minecraftforge.client.gui.overlay;
+
+import net.minecraft.resources.ResourceLocation;
+
+/**
+ * An object representation of an {@link IGuiOverlay overlay} with a name.<p/>
+ * Useful to identify overlays in {@link net.minecraftforge.client.event.RenderGuiOverlayEvent}.
+ * <p>
+ * Users should not be instantiating this themselves. Retrieve from {@link GuiOverlayManager}.
+ */
+public record NamedGuiOverlay(ResourceLocation id, IGuiOverlay overlay)
+{
+}
